<feat><3D:Layers> : Affichage des points geojson
